refactor(rust): 🔄 extract the note engine
Move saved-note operations into an independent engine crate with a
private Library owner. Keep Tauri dispatch and events in the shell.

Preserve IPC contracts and move Rust gates to the workspace root.
Validated 82 engine tests, 14 shell tests, all frontend gates, coverage,
and release compilation. CodeRabbit reported no findings.
