+++ .github/workflows/ci.yml
@@ -108,12 +108,10 @@ jobs:
 
       - name: Check unused Rust dependencies
         if: runner.os == 'Linux'
-        working-directory: src-tauri
         run: cargo machete
 
       - name: Check Rust formatting
         if: runner.os == 'Linux'
-        working-directory: src-tauri
         run: cargo fmt --all -- --check
 
       - name: Setup Node for bindings
@@ -126,13 +124,14 @@ jobs:
         if: runner.os == 'Linux'
         run: pnpm bindings:check
 
+      - name: Test the standalone engine
+        run: cargo test -p notras-core --locked
+
       - name: Clippy
-        working-directory: src-tauri
-        run: cargo clippy --locked --all-targets -- -D warnings
+        run: cargo clippy --workspace --locked --all-targets -- -D warnings
 
       - name: 🧪 Rust Tests
-        working-directory: src-tauri
-        run: cargo test --locked
+        run: cargo test --workspace --locked
 
       - name: Rust coverage report
         id: coverage
@@ -144,7 +143,7 @@ jobs:
         uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
         with:
           name: rust-coverage
-          path: src-tauri/target/coverage/
+          path: target/coverage/
           if-no-files-found: error
 
   actions:
