+++ Cargo.toml
@@ -1,0 +1,18 @@
+[workspace]
+members = ["crates/notras-core", "src-tauri"]
+resolver = "2"
+
+[workspace.dependencies]
+log = "0.4"
+rusqlite = { version = "0.37", features = ["bundled"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+specta = { version = "=2.0.0-rc.22", features = ["derive", "serde_json"] }
+tempfile = "3"
+
+[profile.release]
+codegen-units = 1
+lto = true
+opt-level = "s"
+panic = "abort"
+strip = true
